src/data/gee_export.py
"""Google Earth Engine export stub with graceful degradation.

Works without GEE installed or authenticated - provides setup instructions
and availability checks.
"""

import logging
from typing import Any

logger = logging.getLogger("lccomparison.gee_export")

# Graceful import
try:
    import ee
    _EE_AVAILABLE = True
except ImportError:
    ee = None
    _EE_AVAILABLE = False

try:
    import geemap
    _GEEMAP_AVAILABLE = True
except ImportError:
    geemap = None
    _GEEMAP_AVAILABLE = False


class GEEExporter:
    """Google Earth Engine imagery exporter.

    Provides graceful degradation when GEE is not installed or authenticated.
    """

    def __init__(self, config: dict[str, Any] | None = None):
        self.config = config or {}
        self._authenticated = False
        self._check_availability()

    def _check_availability(self) -> None:
        """Check if GEE is available and authenticated."""
        if not _EE_AVAILABLE:
            logger.warning("earthengine-api not installed. GEE export unavailable.")
            return

        try:
            ee.Initialize()
            self._authenticated = True
            logger.info("GEE authenticated successfully.")
        except Exception as e:
            logger.warning(f"GEE authentication failed: {e}")
            self._authenticated = False

    def is_available(self) -> bool:
        """Check if GEE export is available (installed + authenticated)."""
        return _EE_AVAILABLE and self._authenticated

    def export_sentinel2(
        self,
        bbox: dict[str, float],
        date_range: tuple[str, str],
        output_path: str,
        scale: int = 10,
        bands: list[str] | None = None,
        cloud_filter: int = 20,
    ) -> dict[str, Any]:
        """Export Sentinel-2 imagery for a bounding box.

        Args:
            bbox: Dict with west, south, east, north keys.
            date_range: Tuple of (start_date, end_date) strings.
            output_path: Path for output file.
            scale: Export resolution in meters.
            bands: Band names to export.
            cloud_filter: Maximum cloud cover percentage.

        Returns:
            Dict with export task info.

        Raises:
            RuntimeError: If GEE is not available.
        """
        if not self.is_available():
            raise RuntimeError(
                "GEE is not available. " + self.get_setup_instructions()
            )

        # Stub - full implementation in Phase 2
        logger.info(
            f"Sentinel-2 export stub called: bbox={bbox}, "
            f"dates={date_range}, scale={scale}"
        )
        return {
            "status": "stub",
            "message": "Full GEE export will be implemented in Phase 2",
            "bbox": bbox,
            "date_range": date_range,
        }

    def export_naip(
        self,
        bbox: dict[str, float],
        year: int = 2023,
        output_path: str = "",
        scale: int = 1,
    ) -> dict[str, Any]:
        """Export NAIP imagery for a bounding box.

        Args:
            bbox: Dict with west, south, east, north keys.
            year: NAIP collection year.
            output_path: Path for output file.
            scale: Export resolution in meters.

        Returns:
            Dict with export task info.

        Raises:
            RuntimeError: If GEE is not available.
        """
        if not self.is_available():
            raise RuntimeError(
                "GEE is not available. " + self.get_setup_instructions()
            )

        logger.info(f"NAIP export stub called: bbox={bbox}, year={year}")
        return {
            "status": "stub",
            "message": "Full GEE export will be implemented in Phase 2",
            "bbox": bbox,
            "year": year,
        }

    def check_export_status(self, task_id: str) -> dict[str, Any]:
        """Check the status of a GEE export task.

        Args:
            task_id: GEE task identifier.

        Returns:
            Dict with task status info.
        """
        if not self.is_available():
            return {"status": "unavailable", "task_id": task_id}

        logger.info(f"Export status check stub: task_id={task_id}")
        return {
            "status": "stub",
            "task_id": task_id,
            "message": "Status checking will be implemented in Phase 2",
        }

    @staticmethod
    def get_setup_instructions() -> str:
        """Get instructions for setting up GEE access."""
        return (
            "To set up Google Earth Engine:\n"
            "1. Install: pip install earthengine-api geemap\n"
            "2. Authenticate: earthengine authenticate\n"
            "3. Or use a service account with ee.ServiceAccountCredentials\n"
            "4. Verify: python -c \"import ee; ee.Initialize(); print('OK')\"\n"
        )
