pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "lccomparison"
version = "0.1.0"
description = "Land Cover Model Comparison Framework - Compare foundation models against existing LCAnalysis2026 models"
readme = "README.md"
requires-python = ">=3.10"
license = {text = "MIT"}

dependencies = [
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "rasterio>=1.3.0",
    "geopandas>=0.14.0",
    "shapely>=2.0.0",
    "pyyaml>=6.0",
    "omegaconf>=2.3.0",
    "click>=8.0.0",
    "tqdm>=4.65.0",
    "rich>=13.0.0",
    "matplotlib>=3.7.0",
    "plotly>=5.0.0",
    "xgboost>=2.0.0",
    "rasterstats>=0.19.0",
    "torch>=2.0.0",
    "earthengine-api>=0.1.370",
    "geemap>=0.30.0",
    "google-cloud-storage>=2.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "black>=23.0.0",
    "ruff>=0.0.270",
]

[project.scripts]
lccompare = "src.pipeline:cli"

[tool.setuptools.packages.find]
where = ["."]
include = ["src*", "config*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.ruff]
line-length = 100
select = ["E", "F", "I", "N", "W"]
ignore = ["E501"]
